[update] 攻撃廻り更新

+++ Assets/Script/Enemy/Structure/EnemyStatus.cs
@@ -1,5 +1,5 @@
 [System.Serializable]
 public struct EnemyStatus
 {
-    public int _life;
+    public float _life;
 }
